refactor(api): project_configs 훅·커맨드·메모리·규칙 라우트를 전용 모듈로 추출
24개 라우트 이동 (hooks 5 · commands 6 · memories 7 · rules 6). AST 이름 기반
추출이라 핸들러 본문·경로·함수명 무변경.

memories 는 **모듈 안**의 선언 순서가 계약이다 —
PUT /{project_id}/memories/{memory_id} 가 /{memory_id} 자리에 리터럴 index 를
받아 PUT /{project_id}/memories/index 를 삼킨다. include 순서로는 못 풀어서
원본 선언 순서(index 핸들러 먼저)를 그대로 유지하고 독스트링에 명시했다.

rules 는 __init__.py 에서 전역 설정보다 뒤에 온다 — {project_id}='global' 이면
/global/rules 계열 4개를 통째로 삼킨다.

R2b(split_audit) 통과: 정의 61개 전부 바이트 동일. 라우트 테이블 60개 동일.
