+++ src/backend/api/project_configs/_legacy.py
@@ -13,21 +13,8 @@ from fastapi.responses import StreamingResponse
 
 from api.deps import get_current_user_optional
 from models.project_config import (
-    CommandConfig,
-    CommandContentResponse,
-    CommandCreateRequest,
-    CommandUpdateRequest,
-    CopyCommandRequest,
-    CopyHookRequest,
-    CopyRuleRequest,
     ExternalPathRequest,
     GlobalConfigSummary,
-    HookEntryRequest,
-    HooksUpdateRequest,
-    MemoryConfig,
-    MemoryContentResponse,
-    MemoryCreateRequest,
-    MemoryUpdateRequest,
     ProjectConfigResponse,
     ProjectConfigSummary,
     RuleConfig,
@@ -575,241 +562,9 @@ async def get_project_summary(project_id: str) -> ProjectConfigSummary:
 # ========================================
 
 
-@router.get("/{project_id}/hooks")
-async def list_project_hooks(project_id: str) -> dict:
-    """Get hooks configuration for a project.
-
-    Args:
-        project_id: Project identifier
-
-    Returns:
-        List of hook configurations
-    """
-    monitor = get_project_config_monitor()
-    hooks = monitor.get_project_hooks(project_id)
-
-    return {
-        "project_id": project_id,
-        "hooks": [h.model_dump() for h in hooks],
-        "hook_count": len(hooks),
-    }
-
-
-@router.put("/{project_id}/hooks")
-async def update_hooks(project_id: str, request: HooksUpdateRequest) -> dict:
-    """Update entire hooks.json content.
-
-    Args:
-        project_id: Project identifier
-        request: Complete hooks configuration
-
-    Returns:
-        Success status
-    """
-    monitor = get_project_config_monitor()
-
-    if monitor.update_hooks(project_id, {"hooks": request.hooks}):
-        audit_config_change(AuditAction.CONFIG_UPDATED, "hooks", "hooks.json", project_id)
-        return {
-            "success": True,
-            "message": "Updated hooks configuration",
-            "project_id": project_id,
-        }
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail="Failed to update hooks. Check if project exists.",
-        )
-
-
-@router.post("/{project_id}/hooks/events/{event}")
-async def add_hook_entry(project_id: str, event: str, request: HookEntryRequest) -> dict:
-    """Add a hook entry to an event.
-
-    Args:
-        project_id: Project identifier
-        event: Event name (PreToolUse, PostToolUse, etc.)
-        request: Hook entry request
-
-    Returns:
-        Success status
-    """
-    monitor = get_project_config_monitor()
-
-    if monitor.add_hook_entry(project_id, event, request.matcher, request.hooks):
-        audit_config_change(AuditAction.CONFIG_CREATED, "hook", event, project_id)
-        return {
-            "success": True,
-            "message": f"Added hook entry for event: {event}",
-            "project_id": project_id,
-            "event": event,
-        }
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to add hook entry for event: {event}",
-        )
-
-
-@router.delete("/{project_id}/hooks/{event}/{index}")
-async def delete_hook(project_id: str, event: str, index: int) -> dict:
-    """Delete a hook entry by event and index.
-
-    Args:
-        project_id: Project identifier
-        event: Event name
-        index: Index of hook entry within the event
-
-    Returns:
-        Success status
-    """
-    monitor = get_project_config_monitor()
-
-    if monitor.delete_hook(project_id, event, index):
-        audit_config_change(AuditAction.CONFIG_DELETED, "hook", f"{event}[{index}]", project_id)
-        return {
-            "success": True,
-            "message": f"Deleted hook {event}[{index}]",
-            "project_id": project_id,
-            "event": event,
-            "index": index,
-        }
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to delete hook {event}[{index}]. Check if project and hook exist.",
-        )
-
-
 # ========================================
 # Commands
 # ========================================
-
-
-@router.get("/{project_id}/commands", response_model=list[CommandConfig])
-async def list_project_commands(project_id: str) -> list[CommandConfig]:
-    """Get all commands for a specific project.
-
-    Args:
-        project_id: Project identifier
-
-    Returns:
-        List of commands for the project
-    """
-    monitor = get_project_config_monitor()
-    commands = monitor.get_project_commands(project_id)
-
-    if not commands:
-        # Check if project exists
-        summary = monitor.get_project_summary(project_id)
-        if summary is None:
-            raise HTTPException(status_code=404, detail=f"Project not found: {project_id}")
-
-    return commands
-
-
-@router.get("/{project_id}/commands/{command_id}/content", response_model=CommandContentResponse)
-async def get_command_content(project_id: str, command_id: str) -> CommandContentResponse:
-    """Get full content of a command.
-
-    Args:
-        project_id: Project identifier
-        command_id: Command identifier
-
-    Returns:
-        Command configuration with full content
-    """
-    monitor = get_project_config_monitor()
-    command, content = monitor.get_command_content(project_id, command_id)
-
-    if command is None:
-        raise HTTPException(
-            status_code=404,
-            detail=f"Command not found: {command_id} in project {project_id}",
-        )
-
-    return CommandContentResponse(command=command, content=content)
-
-
-@router.post("/{project_id}/commands", response_model=CommandConfig)
-async def create_command(project_id: str, request: CommandCreateRequest) -> CommandConfig:
-    """Create a new command.
-
-    Args:
-        project_id: Project identifier
-        request: Command create request
-
-    Returns:
-        Created command configuration
-    """
-    monitor = get_project_config_monitor()
-    result = monitor.create_command(project_id, request.command_id, request.content)
-
-    if result is None:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to create command: {request.command_id}. Check if project exists and command ID is unique.",
-        )
-
-    audit_config_change(AuditAction.CONFIG_CREATED, "command", request.command_id, project_id)
-    return result
-
-
-@router.put("/{project_id}/commands/{command_id}")
-async def update_command(project_id: str, command_id: str, request: CommandUpdateRequest) -> dict:
-    """Update command content.
-
-    Args:
-        project_id: Project identifier
-        command_id: Command identifier
-        request: Update request with new content
-
-    Returns:
-        Success status
-    """
-    monitor = get_project_config_monitor()
-
-    if monitor.update_command_content(project_id, command_id, request.content):
-        audit_config_change(AuditAction.CONFIG_UPDATED, "command", command_id, project_id)
-        return {
-            "success": True,
-            "message": f"Updated command: {command_id}",
-            "project_id": project_id,
-            "command_id": command_id,
-        }
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to update command: {command_id}. Check if project and command exist.",
-        )
-
-
-@router.delete("/{project_id}/commands/{command_id}")
-async def delete_command(project_id: str, command_id: str) -> dict:
-    """Delete a command.
-
-    Args:
-        project_id: Project identifier
-        command_id: Command identifier
-
-    Returns:
-        Success status
-    """
-    monitor = get_project_config_monitor()
-
-    if monitor.delete_command(project_id, command_id):
-        audit_config_change(AuditAction.CONFIG_DELETED, "command", command_id, project_id)
-        return {
-            "success": True,
-            "message": f"Deleted command: {command_id}",
-            "project_id": project_id,
-            "command_id": command_id,
-        }
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to delete command: {command_id}. Check if project and command exist.",
-        )
 
 
 # ========================================
@@ -817,402 +572,11 @@ async def delete_command(project_id: str, command_id: str) -> dict:
 # ========================================
 
 
-@router.get("/{project_id}/memories", response_model=list[MemoryConfig])
-async def list_project_memories(project_id: str) -> list[MemoryConfig]:
-    """Get all memory entries for a specific project.
-
-    Args:
-        project_id: Project identifier
-
-    Returns:
-        List of memory entries for the project
-    """
-    monitor = get_project_config_monitor()
-    memories = monitor.get_project_memories(project_id)
-
-    if not memories:
-        summary = monitor.get_project_summary(project_id)
-        if summary is None:
-            raise HTTPException(status_code=404, detail=f"Project not found: {project_id}")
-
-    return memories
-
-
-@router.get("/{project_id}/memories/index")
-async def get_memory_index(project_id: str) -> dict:
-    """Get the MEMORY.md index content for a project.
-
-    Args:
-        project_id: Project identifier
-
-    Returns:
-        Dict with MEMORY.md content
-    """
-    monitor = get_project_config_monitor()
-    content = monitor.get_memory_index(project_id)
-
-    return {
-        "project_id": project_id,
-        "content": content,
-    }
-
-
-@router.put("/{project_id}/memories/index")
-async def update_memory_index(project_id: str, request: MemoryUpdateRequest) -> dict:
-    """Update the MEMORY.md index content for a project.
-
-    Args:
-        project_id: Project identifier
-        request: Update request with new content
-
-    Returns:
-        Success status
-    """
-    monitor = get_project_config_monitor()
-
-    if monitor.update_memory_index(project_id, request.content):
-        audit_config_change(AuditAction.CONFIG_UPDATED, "memory", "MEMORY.md", project_id)
-        return {
-            "success": True,
-            "message": "Updated MEMORY.md index",
-            "project_id": project_id,
-        }
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail="Failed to update MEMORY.md. Check if project exists.",
-        )
-
-
-@router.get("/{project_id}/memories/{memory_id}/content", response_model=MemoryContentResponse)
-async def get_memory_content(project_id: str, memory_id: str) -> MemoryContentResponse:
-    """Get full content of a memory entry.
-
-    Args:
-        project_id: Project identifier
-        memory_id: Memory identifier
-
-    Returns:
-        Memory configuration with full content
-    """
-    monitor = get_project_config_monitor()
-    memory, content = monitor.get_memory_content(project_id, memory_id)
-
-    if memory is None:
-        raise HTTPException(
-            status_code=404,
-            detail=f"Memory not found: {memory_id} in project {project_id}",
-        )
-
-    return MemoryContentResponse(memory=memory, content=content)
-
-
-@router.post("/{project_id}/memories", response_model=MemoryConfig)
-async def create_memory(project_id: str, request: MemoryCreateRequest) -> MemoryConfig:
-    """Create a new memory entry.
-
-    Args:
-        project_id: Project identifier
-        request: Memory create request
-
-    Returns:
-        Created memory configuration
-    """
-    monitor = get_project_config_monitor()
-    result = monitor.create_memory(project_id, request.memory_id, request.content)
-
-    if result is None:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to create memory: {request.memory_id}. Check if project exists and memory ID is unique.",
-        )
-
-    audit_config_change(AuditAction.CONFIG_CREATED, "memory", request.memory_id, project_id)
-    return result
-
-
-@router.put("/{project_id}/memories/{memory_id}")
-async def update_memory(project_id: str, memory_id: str, request: MemoryUpdateRequest) -> dict:
-    """Update memory content.
-
-    Args:
-        project_id: Project identifier
-        memory_id: Memory identifier
-        request: Update request with new content
-
-    Returns:
-        Success status
-    """
-    monitor = get_project_config_monitor()
-
-    if monitor.update_memory_content(project_id, memory_id, request.content):
-        audit_config_change(AuditAction.CONFIG_UPDATED, "memory", memory_id, project_id)
-        return {
-            "success": True,
-            "message": f"Updated memory: {memory_id}",
-            "project_id": project_id,
-            "memory_id": memory_id,
-        }
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to update memory: {memory_id}. Check if project and memory exist.",
-        )
-
-
-@router.delete("/{project_id}/memories/{memory_id}")
-async def delete_memory(project_id: str, memory_id: str) -> dict:
-    """Delete a memory entry.
-
-    Args:
-        project_id: Project identifier
-        memory_id: Memory identifier
-
-    Returns:
-        Success status
-    """
-    monitor = get_project_config_monitor()
-
-    if monitor.delete_memory(project_id, memory_id):
-        audit_config_change(AuditAction.CONFIG_DELETED, "memory", memory_id, project_id)
-        return {
-            "success": True,
-            "message": f"Deleted memory: {memory_id}",
-            "project_id": project_id,
-            "memory_id": memory_id,
-        }
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to delete memory: {memory_id}. Check if project and memory exist.",
-        )
-
-
 # ========================================
 # Rules
 # ========================================
 
 
-@router.get("/{project_id}/rules", response_model=list[RuleConfig])
-async def list_project_rules(project_id: str) -> list[RuleConfig]:
-    """Get all rules for a specific project.
-
-    Args:
-        project_id: Project identifier
-
-    Returns:
-        List of rules for the project
-    """
-    monitor = get_project_config_monitor()
-    rules = monitor.get_project_rules(project_id)
-
-    if not rules:
-        summary = monitor.get_project_summary(project_id)
-        if summary is None:
-            raise HTTPException(status_code=404, detail=f"Project not found: {project_id}")
-
-    return rules
-
-
-@router.get("/{project_id}/rules/{rule_id}/content", response_model=RuleContentResponse)
-async def get_rule_content(project_id: str, rule_id: str) -> RuleContentResponse:
-    """Get full content of a project rule.
-
-    Args:
-        project_id: Project identifier
-        rule_id: Rule identifier
-
-    Returns:
-        Rule configuration with full content
-    """
-    monitor = get_project_config_monitor()
-    rule, content = monitor.get_rule_content(project_id, rule_id, is_global=False)
-
-    if rule is None:
-        raise HTTPException(
-            status_code=404,
-            detail=f"Rule not found: {rule_id} in project {project_id}",
-        )
-
-    return RuleContentResponse(rule=rule, content=content)
-
-
-@router.post("/{project_id}/rules", response_model=RuleConfig)
-async def create_rule(project_id: str, request: RuleCreateRequest) -> RuleConfig:
-    """Create a new project rule.
-
-    Args:
-        project_id: Project identifier
-        request: Rule create request
-
-    Returns:
-        Created rule configuration
-    """
-    monitor = get_project_config_monitor()
-    result = monitor.create_rule(project_id, request.rule_id, request.content, is_global=False)
-
-    if result is None:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to create rule: {request.rule_id}. Check if project exists and rule ID is unique.",
-        )
-
-    audit_config_change(AuditAction.CONFIG_CREATED, "rules", request.rule_id, project_id)
-    return result
-
-
-@router.put("/{project_id}/rules/{rule_id}")
-async def update_rule(project_id: str, rule_id: str, request: RuleUpdateRequest) -> dict:
-    """Update project rule content.
-
-    Args:
-        project_id: Project identifier
-        rule_id: Rule identifier
-        request: Update request with new content
-
-    Returns:
-        Success status
-    """
-    monitor = get_project_config_monitor()
-
-    if monitor.update_rule_content(project_id, rule_id, request.content, is_global=False):
-        audit_config_change(AuditAction.CONFIG_UPDATED, "rules", rule_id, project_id)
-        return {
-            "success": True,
-            "message": f"Updated rule: {rule_id}",
-            "project_id": project_id,
-            "rule_id": rule_id,
-        }
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to update rule: {rule_id}. Check if project and rule exist.",
-        )
-
-
-@router.delete("/{project_id}/rules/{rule_id}")
-async def delete_rule(project_id: str, rule_id: str) -> dict:
-    """Delete a project rule.
-
-    Args:
-        project_id: Project identifier
-        rule_id: Rule identifier
-
-    Returns:
-        Success status
-    """
-    monitor = get_project_config_monitor()
-
-    if monitor.delete_rule(project_id, rule_id, is_global=False):
-        audit_config_change(AuditAction.CONFIG_DELETED, "rules", rule_id, project_id)
-        return {
-            "success": True,
-            "message": f"Deleted rule: {rule_id}",
-            "project_id": project_id,
-            "rule_id": rule_id,
-        }
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to delete rule: {rule_id}. Check if project and rule exist.",
-        )
-
-
-@router.post("/{project_id}/rules/{rule_id}/copy")
-async def copy_rule(project_id: str, rule_id: str, request: CopyRuleRequest) -> dict:
-    """Copy a rule to another project.
-
-    Args:
-        project_id: Source project identifier
-        rule_id: Rule identifier to copy
-        request: Copy request with target project
-
-    Returns:
-        Success status
-    """
-    monitor = get_project_config_monitor()
-
-    result = monitor.copy_rule(project_id, rule_id, request.target_project_id)
-
-    if result:
-        return {
-            "success": True,
-            "message": f"Copied rule '{rule_id}' to project '{request.target_project_id}'",
-            "source_project_id": project_id,
-            "target_project_id": request.target_project_id,
-            "rule_id": rule_id,
-        }
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to copy rule '{rule_id}'. Check if source and target projects exist.",
-        )
-
-
 # ========================================
 # Copy Operations
 # ========================================
-
-
-@router.post("/{project_id}/hooks/{event}/{index}/copy")
-async def copy_hook(project_id: str, event: str, index: int, request: CopyHookRequest) -> dict:
-    """Copy a hook to another project.
-
-    Args:
-        project_id: Source project identifier
-        event: Hook event name
-        index: Hook entry index
-        request: Copy request with target project
-
-    Returns:
-        Success status
-    """
-    monitor = get_project_config_monitor()
-
-    result = monitor.copy_hook(project_id, event, index, request.target_project_id)
-
-    if result:
-        return {
-            "success": True,
-            "message": f"Copied hook '{event}[{index}]' to project '{request.target_project_id}'",
-            "source_project_id": project_id,
-            "target_project_id": request.target_project_id,
-            "event": event,
-            "index": index,
-        }
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to copy hook '{event}[{index}]'. Check if source and target projects exist.",
-        )
-
-
-@router.post("/{project_id}/commands/{command_id}/copy")
-async def copy_command(project_id: str, command_id: str, request: CopyCommandRequest) -> dict:
-    """Copy a command to another project.
-
-    Args:
-        project_id: Source project identifier
-        command_id: Command identifier to copy
-        request: Copy request with target project
-
-    Returns:
-        Success status
-    """
-    monitor = get_project_config_monitor()
-
-    result = monitor.copy_command(project_id, command_id, request.target_project_id)
-
-    if result:
-        return {
-            "success": True,
-            "message": f"Copied command '{command_id}' to project '{request.target_project_id}'",
-            "source_project_id": project_id,
-            "target_project_id": request.target_project_id,
-            "command_id": command_id,
-        }
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Failed to copy command '{command_id}'. Check if source and target projects exist.",
-        )
